PROGETTO FINALE SIGNORE-FERRAUTO/data_collector/requirements.txt
yfinance
mysql-connector-python
schedule
confluent_kafka
prometheus-client
